third_party/WebKit/LayoutTests/http/tests/devtools/console/console-edit-expanded-tree-expected.txt

CONSOLE MESSAGE: line 12: [object Object]
Tests that expanded tree element is editable in console.

After viewport refresh tree element remains in editing mode: true
